Move enum to folder remove comments
