--- Weatherman.Console/Weatherman.Core/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Weatherman.Core
-{
-    public class Class1
-    {
-    }
-}
